+++ src/main/java/ru/practicum/shareit/booking/Booking.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.booking;
+
+/**
+ * TODO Sprint add-bookings.
+ */
+public class Booking {
+}
