+++ src/main/java/ru/practicum/shareit/booking/dto/BookingDto.java
@@ -1,25 +1,7 @@
 package ru.practicum.shareit.booking.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import ru.practicum.shareit.booking.model.BookingStatus;
-import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.user.model.User;
-
-import java.time.LocalDateTime;
-
 /**
  * TODO Sprint add-bookings.
  */
-@Getter
-@Setter
-@AllArgsConstructor
 public class BookingDto {
-    private long id;
-    private LocalDateTime start;
-    private LocalDateTime end;
-    private Item item;
-    private User booker;
-    private BookingStatus status;
 }
