--- src/main/java/ru/practicum/shareit/booking/model/Booking.java
@@ -1,25 +1,0 @@
-package ru.practicum.shareit.booking.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.user.model.User;
-
-import java.time.LocalDateTime;
-
-/**
- * TODO Sprint add-bookings.
- */
-@Getter
-@Setter
-@AllArgsConstructor
-public class Booking {
-    private long id;
-    private LocalDateTime start;
-    private LocalDateTime end;
-    private Item item;
-    private User booker;
-    private BookingStatus status;
-
-}
