--- src/main/java/ru/practicum/shareit/exception/ErrorHandler.java
@@ -1,42 +1,0 @@
-package ru.practicum.shareit.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import javax.validation.ValidationException;
-
-@Slf4j
-@RestControllerAdvice
-public class ErrorHandler {
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleValidationException(final ValidationException e) {
-        log.error("Ошибка валидации, {}", e.getMessage());
-        return new ErrorResponse(
-                e.getMessage()
-        );
-    }
-
-    @ExceptionHandler
-    @ResponseStatus (HttpStatus.BAD_REQUEST)
-    public ErrorResponse validateArgumentException(final MethodArgumentNotValidException e) {
-        log.error("Validation error, {}", e.getMessage());
-        return new ErrorResponse(
-                e.getMessage()
-        );
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse handleNotFoundException(final NotFoundException e) {
-        log.error("NotFoundException, {}", e.getMessage());
-        return new ErrorResponse(
-                e.getMessage()
-        );
-    }
-}
