--- src/main/java/ru/practicum/shareit/exception/NotFoundException.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class NotFoundException extends RuntimeException {
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
