--- src/main/java/ru/practicum/shareit/exception/ValidationException.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class ValidationException extends RuntimeException {
-
-    public ValidationException(String message) {
-        super(message);
-    }
-}
