+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.item;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * TODO Sprint add-controllers.
+ */
+@RestController
+@RequestMapping("/items")
+public class ItemController {
+}
