--- src/main/java/ru/practicum/shareit/item/controller/ItemController.java
@@ -1,57 +1,0 @@
-package ru.practicum.shareit.item.controller;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.item.dto.ItemDto;
-import ru.practicum.shareit.item.model.ItemMapper;
-import ru.practicum.shareit.item.service.ItemService;
-
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-import java.util.List;
-import java.util.Map;
-
-/**
- * TODO Sprint add-controllers.
- */
-
-@Slf4j
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/items")
-public class ItemController {
-    private final ItemService itemService;
-
-    @GetMapping("/{itemId}")
-    public ItemDto get(@PathVariable long itemId) {
-        return itemService.get(itemId);
-    }
-
-    @GetMapping
-    public List<ItemDto> getAllItemsUser(@RequestHeader("X-Sharer-User-Id") long ownerId) {
-        return itemService.getAllItemsUser(ownerId);
-    }
-
-    @PostMapping
-    public ItemDto create(@RequestHeader("X-Sharer-User-Id") long ownerId, @Valid @RequestBody @NotNull ItemDto itemDto) {
-        return itemService.create(ownerId, ItemMapper.fromItem(itemDto));
-    }
-
-    @PatchMapping("/{itemId}")
-    public ItemDto patch(@RequestHeader("X-Sharer-User-Id") long ownerId, @PathVariable long itemId,
-                         @RequestBody @NotNull Map<String, String> updates) {
-        return itemService.update(ownerId, itemId, updates);
-    }
-
-    @GetMapping("/search")
-    public List<ItemDto> search(@RequestParam(name = "text") String text) {
-        return itemService.search(text);
-    }
-
-    @DeleteMapping("/{itemId}")
-    public void delete(@PathVariable long itemId) {
-        itemService.delete(itemId);
-    }
-
-}
