+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -4,5 +4,4 @@ package ru.practicum.shareit.item.dto;
  * TODO Sprint add-controllers.
  */
 public class ItemDto {
-    
 }
