+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -1,33 +1,7 @@
 package ru.practicum.shareit.item.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-import ru.practicum.shareit.user.model.User;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
-
 /**
  * TODO Sprint add-controllers.
  */
-@Builder
-@Getter
-@Setter
-@AllArgsConstructor
 public class Item {
-    private long id;
-    @NotBlank
-    @Size(max = 50)
-    private String name;
-    @NotBlank
-    @Size(max = 200)
-    private String description;
-    @NotNull
-    private Boolean available;
-    @NotNull
-    private User owner;
-    private Long request;
 }
