--- src/main/java/ru/practicum/shareit/item/model/ItemMapper.java
@@ -1,36 +1,0 @@
-package ru.practicum.shareit.item.model;
-
-import ru.practicum.shareit.item.dto.ItemDto;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class ItemMapper {
-    public static ItemDto toItemDto(Item item) {
-        return ItemDto.builder()
-                .id(item.getId())
-                .name(item.getName())
-                .description(item.getDescription())
-                .available(item.getAvailable())
-                .owner(item.getOwner())
-                .request(item.getRequest())
-                .build();
-    }
-
-    public static Item fromItem(ItemDto itemDto) {
-        return Item.builder()
-                .id(itemDto.getId())
-                .name(itemDto.getName())
-                .description(itemDto.getDescription())
-                .available(itemDto.getAvailable())
-                .owner(itemDto.getOwner())
-                .request(itemDto.getRequest())
-                .build();
-    }
-
-    public static List<ItemDto> toItemDtoList(List<Item> items) {
-        return items.stream()
-                .map(ItemMapper::toItemDto)
-                .collect(Collectors.toList());
-    }
-}
