+++ src/main/java/ru/practicum/shareit/request/ItemRequest.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.request;
+
+/**
+ * TODO Sprint add-item-requests.
+ */
+public class ItemRequest {
+}
