+++ src/main/java/ru/practicum/shareit/request/dto/ItemRequestDto.java
@@ -1,12 +1,7 @@
 package ru.practicum.shareit.request.dto;
 
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
 /**
  * TODO Sprint add-item-requests.
  */
-@RestController
-@RequestMapping(path = "/requests")
 public class ItemRequestDto {
 }
