--- src/main/java/ru/practicum/shareit/request/model/ItemRequest.java
@@ -1,20 +1,0 @@
-package ru.practicum.shareit.request.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import ru.practicum.shareit.user.model.User;
-
-import java.time.LocalDateTime;
-
-/**
- * TODO Sprint add-item-requests.
- */
-
-@Data
-@AllArgsConstructor
-public class ItemRequest {
-    private long id;
-    private String description;
-    private User requestor;
-    private LocalDateTime created;
-}
