+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -1,13 +1,12 @@
-package ru.practicum.shareit.booking.controller;
+package ru.practicum.shareit.user;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
- * TODO Sprint add-bookings.
+ * TODO Sprint add-controllers.
  */
 @RestController
-@RequestMapping(path = "/bookings")
-public class BookingController {
-
+@RequestMapping(path = "/users")
+public class UserController {
 }
