--- src/main/java/ru/practicum/shareit/user/controller/UserController.java
@@ -1,44 +1,0 @@
-package ru.practicum.shareit.user.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.model.UserMapper;
-import ru.practicum.shareit.user.service.UserService;
-
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-import java.util.List;
-import java.util.Map;
-
-@RestController
-@RequestMapping(path = "/users")
-@RequiredArgsConstructor
-public class UserController {
-    private final UserService userService;
-
-    @GetMapping("{id}")
-    public UserDto getUser(@PathVariable(required = false) long id) {
-        return userService.getUser(id);
-    }
-
-    @GetMapping
-    public List<UserDto> getAllUsers() {
-        return userService.getAll();
-    }
-
-    @PostMapping
-    public UserDto create(@Valid @RequestBody @NotNull UserDto user) {
-        return userService.create(UserMapper.fromUser(user));
-    }
-
-    @PatchMapping("{id}")
-    public UserDto patchUpdate(@PathVariable long id, @RequestBody Map<String, String> updates) {
-        return userService.update(id, updates);
-    }
-
-    @DeleteMapping("{id}")
-    public void deleteUser(@PathVariable long id) {
-        userService.delete(id);
-    }
-}
