--- src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -1,22 +1,0 @@
-package ru.practicum.shareit.user.dto;
-
-import lombok.*;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-
-@Builder
-@Getter
-@Setter
-@AllArgsConstructor
-public class UserDto {
-    private long id;
-    @NotBlank
-    @Size(max = 50)
-    private String name;
-    @NotBlank
-    @Email
-    @Size(max = 256)
-    private String email;
-}
