--- src/main/java/ru/practicum/shareit/user/model/UserMapper.java
@@ -1,30 +1,0 @@
-package ru.practicum.shareit.user.model;
-
-import ru.practicum.shareit.user.dto.UserDto;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-public final class UserMapper {
-    public static UserDto toUserDto(User user) {
-        return UserDto.builder()
-                .id(user.getId())
-                .email(user.getEmail())
-                .name(user.getName())
-                .build();
-    }
-
-    public static List<UserDto> toUserDtoList(List<User> users) {
-        return users.stream()
-                .map(UserMapper::toUserDto)
-                .collect(Collectors.toList());
-    }
-
-    public static User fromUser(UserDto user) {
-        return User.builder()
-                .id(user.getId())
-                .email(user.getEmail())
-                .name(user.getName())
-                .build();
-    }
-}
